Group conversations now dynamic
